BugSearch/app/requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.2
aniso8601==1.1.0
elasticsearch==2.1.0
elasticsearch-dsl==0.0.9
itsdangerous==0.24
python-dateutil==2.4.2
pytz==2015.7
simplejson==3.8.1
six==1.10.0
urllib3==1.12